requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
openai==1.3.0